- removed a few unused parameters, unused casts, unused nullchecks, ...
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@6378 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/SplashScreen.java
@@ -120,8 +120,8 @@ public class SplashScreen implements ActionListener, ControlListener {
         Rectangle v = ge.getDefaultScreenDevice().getDefaultConfiguration().getBounds();
         int screenWidth = (int) v.getWidth();
         int screenHeight = (int) v.getHeight();
-        x = (int) (screenWidth / 2 - image.getWidth(null) / 2);
-        y = (int) (screenHeight / 2 - image.getHeight(null) / 2);
+        x = screenWidth / 2 - image.getWidth(null) / 2;
+        y = screenHeight / 2 - image.getHeight(null) / 2;
         w = image.getWidth(null);
         h = image.getHeight(null);
         initGui();
